contest/compose/highfreshness/client/requirements.txt
altair==4.2.0
attrs==22.1.0
# backports.zoneinfo==0.2.1
backports.zoneinfo;python_version<"3.9"
blinker==1.5
cachetools==5.2.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
decorator==5.1.1
entrypoints==0.4
gitdb==4.0.10
GitPython==3.1.29
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.1
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
MouseInfo==0.1.3
numpy==1.23.5
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
pyarrow==10.0.1
PyAutoGUI==0.9.53
pydeck==0.8.0
PyGetWindow==0.0.9
Pygments==2.13.0
Pympler==1.0.1
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.19.2
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
qrcode==7.3.1
requests==2.28.1
rich==12.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.15.2
streamlit-webcam-example==0.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
watchdog==2.2.0
wincertstore==0.2
zipp==3.11.0
